refactor(SingleSelect): move SingleSelect to src root and update import paths
