+++ packages/components/src/Filter/FilterSelect/FilterSelect.tsx
@@ -4,7 +4,7 @@ import { HiddenSelect, useSelect } from '@react-aria/select'
 import { useSelectState, type SelectProps as AriaSelectProps } from '@react-stately/select'
 import { Filter, FilterContents } from '~components/Filter/Filter'
 import { type FilterButtonProps } from '~components/Filter/FilterButton'
-import { SelectProvider } from '~components/__next__/Select/context'
+import { SelectProvider } from '~components/SingleSelect/context'
 import {
   ListBoxSection,
   ListItem,
@@ -12,9 +12,9 @@ import {
   SectionDivider,
   SelectPopoverContents,
   type SelectPopoverContentsProps,
-} from '~components/__next__/Select/subcomponents'
-import { getDisabledKeysFromItems } from '~components/__next__/Select/utils/getDisabledKeysFromItems'
-import { transformSelectItemToCollectionElement } from '~components/__next__/Select/utils/transformSelectItemToCollectionElement'
+} from '~components/SingleSelect/subcomponents'
+import { getDisabledKeysFromItems } from '~components/SingleSelect/utils/getDisabledKeysFromItems'
+import { transformSelectItemToCollectionElement } from '~components/SingleSelect/utils/transformSelectItemToCollectionElement'
 import { type OverrideClassName } from '~components/types/OverrideClassName'
 import { type SelectItem, type SelectOption } from './types'
 import styles from './FilterSelect.module.css'
