+++ packages/components/src/Filter/FilterSelect/types.ts
@@ -1,3 +1,3 @@
 // @todo: Remove this and use the types from Select when it is moved out from next.
 // This is currently here to allow consumers access to the types from the entrypoint.
-export * from '~components/__next__/Select/types'
+export * from '~components/SingleSelect/types'
