+++ packages/components/src/SingleSelect/_docs/Select.mdx
@@ -4,7 +4,7 @@ import * as SelectStories from './Select.stories'
 
 <Meta of={SelectStories} />
 
-# Select (next)
+# SingleSelect
 
 <ResourceLinks
   sourceCode="https://github.com/cultureamp/kaizen-design-system/tree/main/packages/components/src/__next__/Select"
