+++ packages/components/src/SingleSelect/_docs/Select.stickersheet.stories.tsx
@@ -14,7 +14,7 @@ import {
 const IS_CHROMATIC = isChromatic()
 
 export default {
-  title: 'Components/Select/SingleSelect (next)',
+  title: 'Components/SingleSelect',
   parameters: {
     chromatic: { disable: false },
     controls: { disable: true },
