+++ packages/components/src/SingleSelect/index.ts
@@ -1,0 +1,4 @@
+export * from './SingleSelect'
+export * from './types'
+
+export { SingleSelect, type SingleSelectProps } from './SingleSelect'
