+++ packages/components/src/SingleSelect/subcomponents/ListBoxSection/ListBoxSection.module.scss
