+++ packages/components/src/SingleSelect/subcomponents/SectionDivider/SectionDivider.module.scss
