+++ packages/components/src/SingleSelect/subcomponents/SectionDivider/SectionDivider.tsx
