+++ packages/components/src/SingleSelect/subcomponents/SectionDivider/index.ts
