+++ packages/components/src/SingleSelect/subcomponents/SelectToggle/SelectToggle.module.scss
@@ -2,7 +2,7 @@
 @import '~@kaizen/design-tokens/sass/border';
 @import '~@kaizen/design-tokens/sass/spacing';
 @import '~@kaizen/design-tokens/sass/typography';
-@import '../../../../../styles/utils/button-reset';
+@import '../../../../styles/utils/button-reset';
 
 @layer kz-components {
   .iconWrapper {
