+++ packages/components/src/SingleSelect/utils/isSelectOptionGroup.ts
