+++ packages/components/src/SingleSelect/utils/transformSelectItemToCollectionElement.tsx
