--- packages/components/src/__next__/SingleSelect/index.ts
@@ -1,5 +1,0 @@
-export * from './SingleSelect'
-export * from './types'
-
-// Backward compatibility - re-export SingleSelect as Select
-export { SingleSelect as Select, type SingleSelectProps as SelectProps } from './SingleSelect'
