+++ packages/components/src/__next__/index.ts
@@ -1,7 +1,6 @@
 export * from './Button'
 export * from './Icon'
 export * from './Menu'
-export * from './SingleSelect'
 export * from './Tag'
 export * from './Tabs'
 export * from './Tooltip'
