+++ packages/components/src/index.ts
@@ -64,3 +64,4 @@ export * from './utils'
 export * from './VisuallyHidden'
 export * from './Well'
 export * from './Workflow'
+export * from './SingleSelect'
